+++ drivers/usb/common/buf/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Copyright Nordic Semiconductor ASA
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+zephyr_library_sources_ifdef(CONFIG_UDC_DRIVER usb_buf.c)
