+++ drivers/usb/udc/udc_common.c
@@ -21,36 +21,6 @@
 #endif
 LOG_MODULE_REGISTER(udc, CONFIG_UDC_DRIVER_LOG_LEVEL);
 
-static inline uint8_t *udc_pool_data_alloc(struct net_buf *const buf,
-					   size_t *const size, k_timeout_t timeout)
-{
-	struct net_buf_pool *const buf_pool = net_buf_pool_get(buf->pool_id);
-	struct k_heap *const pool = buf_pool->alloc->alloc_data;
-	void *b;
-
-	*size = ROUND_UP(*size, UDC_BUF_GRANULARITY);
-	b = k_heap_aligned_alloc(pool, UDC_BUF_ALIGN, *size, timeout);
-	if (b == NULL) {
-		*size = 0;
-		return NULL;
-	}
-
-	return b;
-}
-
-static inline void udc_pool_data_unref(struct net_buf *buf, uint8_t *const data)
-{
-	struct net_buf_pool *buf_pool = net_buf_pool_get(buf->pool_id);
-	struct k_heap *pool = buf_pool->alloc->alloc_data;
-
-	k_heap_free(pool, data);
-}
-
-const struct net_buf_data_cb net_buf_dma_cb = {
-	.alloc = udc_pool_data_alloc,
-	.unref = udc_pool_data_unref,
-};
-
 static inline void udc_buf_destroy(struct net_buf *buf);
 
 UDC_BUF_POOL_VAR_DEFINE(udc_ep_pool,
